cleanup and comment

+++ src/animation/mod.rs
@@ -47,6 +47,7 @@ pub struct TextureData {
     pub animation_direction: AnimationDirection,
 }
 
+/// Describes an animation
 #[derive(Deserialize)]
 pub struct AnimationData {
     pub direction: AnimationDirection,
@@ -85,7 +86,7 @@ pub fn animate_sprite_system(
             match &animation.direction {
                 AnimationDirection::None => {}
                 AnimationDirection::Forward => {
-                    sprite.index = (sprite.index as usize + 1) % texture_atlas.textures.len()
+                    sprite.index = (sprite.index + 1) % texture_atlas.textures.len()
                 }
                 AnimationDirection::PingPong(direction) => match direction {
                     PingPongDirection::Forward => {
